protecting edit, delete video routes

+++ controllers/videoController.js
@@ -76,9 +76,12 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (`${video.creator}` !== `${req.user.id}`) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
-    console.log(error);
     res.redirect(routes.home);
   }
 };
@@ -103,9 +106,12 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    // const video = await Video.findById(id);
-    // res.render("deleteVideo", { pageTitle: "Deleting..." }, video);
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (`${video.creator}` !== `${req.user.id}`) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   } finally {
